requirements.txt
pandas==2.3.3
numpy==2.4.2
scikit-learn==1.8.0
faiss-cpu==1.13.2
ollama==0.6.1
streamlit==1.54.0